requirements.txt
numpy==2.2.5
Pillow==11.2.1
scikit_learn==1.6.1
tensorflow==2.18.0
tensorflow_intel==2.18.0
